Refactor dashboard layout and landing page
- Updated the layout to include a navigation link to the dashboard.
- Replaced the previous clusters list page with a new landing page featuring a hero section and a features grid.
- Enhanced user experience with improved styling and navigation elements.
