+++ dashboard/app/layout.tsx
@@ -1,4 +1,5 @@
 import type { Metadata } from 'next';
+import Link from 'next/link';
 import './globals.css';
 
 export const metadata: Metadata = {
@@ -24,10 +25,18 @@ export default function RootLayout({
       <body className="antialiased bg-gray-50">
         <div className="min-h-screen">
           <header className="bg-white shadow-sm border-b">
-            <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-4">
-              <h1 className="text-2xl font-bold text-gray-900">
-                FeedbackAgent Dashboard
-              </h1>
+            <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-4 flex justify-between items-center">
+              <Link href="/" className="text-2xl font-bold text-gray-900 hover:text-gray-700 transition-colors">
+                FeedbackAgent
+              </Link>
+              <nav>
+                <Link 
+                  href="/clusters" 
+                  className="text-sm font-medium text-gray-500 hover:text-gray-900 transition-colors"
+                >
+                  Dashboard
+                </Link>
+              </nav>
             </div>
           </header>
           <main className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-8">
